Move data structure to core
and fix errors with length and append
