+++ src/go/corerepository/repository.go
@@ -1,10 +1,5 @@
 package corerepository
 
-// Any number of Git repositories.
-type Repositories interface {
-	Names() []string
-}
-
 // One Git repository.
 type Repository struct {
 	Name string
