+++ src/go/userepository/list_repositories_query_test.go
@@ -3,7 +3,6 @@ package userepository_test
 import (
 	core "github.com/kkrull/marmot/corerepository"
 	main "github.com/kkrull/marmot/mainfactory"
-	use "github.com/kkrull/marmot/userepository"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
@@ -12,20 +11,19 @@ var _ = Describe("ListRepositoriesQuery", func() {
 	var (
 		factory *main.CommandFactory
 		source  *MockRepositorySource
-		subject *use.ListRepositoriesQuery
 	)
 
 	Describe("#Run", func() {
-		It("returns an empty result set, given a meta repo with no repositories", func() {
-			source = &MockRepositorySource{Names: make([]string, 0)}
+		It("returns all repositories included the source can list", func() {
+			source = &MockRepositorySource{Names: []string{"one", "two"}}
 			factory = &main.CommandFactory{RepositorySource: source}
 
-			subject, _ = factory.ListRepositoriesQuery()
-			repositories, _ := subject.Run()
-			Expect(repositories.Names()).To(BeEmpty())
-		})
+			subject, factoryErr := factory.ListRepositoriesQuery()
+			Expect(subject, factoryErr).NotTo(BeNil())
 
-		It("returns something interesting", Pending, func() {
+			repositories, runErr := subject.Run()
+			Expect(repositories, runErr).NotTo(BeNil())
+			Expect(repositories.Names()).To(ConsistOf("one", "two"))
 		})
 	})
 })
@@ -35,23 +33,10 @@ type MockRepositorySource struct {
 }
 
 func (source *MockRepositorySource) List() (core.Repositories, error) {
-	repositories := make([]core.Repository, len(source.Names))
+	repositories := make([]core.Repository, 0, len(source.Names))
 	for _, name := range source.Names {
 		repositories = append(repositories, core.Repository{Name: name})
 	}
 
-	return &RepositoriesArray{Repositories: repositories}, nil
-}
-
-type RepositoriesArray struct {
-	Repositories []core.Repository
-}
-
-func (array RepositoriesArray) Names() []string {
-	names := make([]string, len(array.Repositories))
-	for _, repository := range array.Repositories {
-		names = append(names, repository.Name)
-	}
-
-	return names
+	return &core.RepositoriesArray{Repositories: repositories}, nil
 }
